pyproject.toml
[project]
name = "optt2i"
version = "0.1.0"
description = "A Python package for optimizing prompts for text-to-image models"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [
    {name = "FBarrca", email = "fbarraganc@alu.comillas.edu"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.8.0",
    "torchvision",
    "torchaudio",
    "diffusers[torch]>=0.35.1",
    "transformers>=4.56.0",
    "nunchaku",
    "opencv-python",
    "numpy<2",
    "onnxruntime-gpu>=1.22.0",
    "timm>=1.0.19",
    "einops>=0.8.1",
    "insightface>=0.7.3",
    "facexlib>=0.3.0",
    "gradio>=5.44.1",
    "Pillow>=10",
    "google-generativeai>=0.8.3",
    "openai>=1.35.0",
    "sentencepiece>=0.2.0",
    "optuna>=3.6.1",
    "python-dotenv>=1.1.1",
    "hatchling",
    "lmstudio>=1.5.0",
    "pygraphviz>=1.14",
]

[project.scripts]
optt2i = "optt2i.main:main"
"optt2i.t2i" = "optt2i.t2i.generate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.pip]
prerelease = "allow"

[tool.uv.sources]
torch = [
    { index = "pytorch-cuda" },
]
torchvision = [
    { index = "pytorch-cuda" },
]
torchaudio = [
    { index = "pytorch-cuda" },
]
pytorch-triton = [{ index = "pytorch-cuda" }]
nunchaku = { url = "https://github.com/nunchaku-tech/nunchaku/releases/download/v1.0.0/nunchaku-1.0.0+torch2.9-cp312-cp312-linux_x86_64.whl" }
#xformers = { git = "https://github.com/johnnynunez/xformers", rev = "patch-1" }

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/nightly/cu128"
explicit = true

[tool.uv]
override-dependencies = ["pytorch-triton==3.4.0+gitf7888497"]
